Refactoring the cityInfo bean
