+++ app/src/main/java/com/backbase/android/weatherapp/CitiesListActivity/CityInfo.java
@@ -51,4 +51,30 @@ public class CityInfo
     {
         this.coordinate = coordinate;
     }
+
+    private class Coordinate
+    {
+        private long lat;
+        private long lon;
+
+        public long getLat()
+        {
+            return lat;
+        }
+
+        public void setLat(long lat)
+        {
+            this.lat = lat;
+        }
+
+        public long getLon()
+        {
+            return lon;
+        }
+
+        public void setLon(long lon)
+        {
+            this.lon = lon;
+        }
+    }
 }
